% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{display.palette}
\alias{display.palette}
\title{Display palette}
\usage{
display.palette(pal = IMD.palette())
}
\arguments{
\item{palis}{a palette default to IMD.palette}
}
\description{
Show the palette for ASRS presentations
}
\examples{
display.palette()
}
\keyword{palette}
